Move popover size regression test to sandbox
